% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mathUtils.R
\name{harmonic_no}
\alias{harmonic_no}
\title{Harmonic Numbers}
\usage{
harmonic_no(N)
}
\arguments{
\item{a}{number N to compute the harmonic number}
}
\value{
a list of N harmonic numbers, where N is the input parameter.
}
\description{
\code{harmonic_no} returns the a list of harmonic number from 1 to N.
}
\details{
The n-th harmonic number is the sum of the reciprocals of the first
 n natural numbers. For more information, refer to the following wikipedia
 link: http://en.wikipedia.org/wiki/Harmonic_number.
}

