% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adhoc.R
\name{nDCG}
\alias{nDCG}
\title{Normalized Discounted Cumulative Gain at k}
\usage{
nDCG(qrels_grades, grades, ranks)
}
\arguments{
\item{qrels_grades}{a numeric vector containing the grades all the judged
documents in the qrels file.}

\item{grades}{a numeric vector of the grades.}

\item{ranks}{a numeric vector containing a set of ranks at which the measure
is calculated.}
}
\description{
\code{AP}  Precision
}

