\name{nERR}
\alias{nERR}
\title{Normalized Expected Reciprocal Rank at k}
\usage{
nERR(qrels_grades, grades, ranks, maxGrade = 4)
}
\arguments{
  \item{qrels_grades}{a numeric vector containing the
  grades all the judged documents in the qrels file.}

  \item{grades}{a numeric vector of the grades.}

  \item{ranks}{a numeric vector containing a set of ranks
  at which the measure is calculated.}
}
\description{
\code{AP} Precision
}

