% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{nNRBP}
\alias{nNRBP}
\title{Normalized Novelty and Rank Biased Precisionf}
\usage{
nNRBP(qrels_grades, grades, rankLimit = 1000, idealMatrix = NULL,
  alpha = 0.5, beta = 0.5)
}
\arguments{
\item{qrels_grades}{a numeric matrix containing the grades all the judged
documents in the qrels file.}

\item{grades}{a numeric matrix of the grades.}

\item{ranks}{a numeric vector containing a set of ranks at which the measure
is calculated.}
}
\description{
\code{AP}  Precision
}

