% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adhoc.R
\name{num_rel}
\alias{num_rel}
\title{Number of Relevant Documents}
\usage{
num_rel(qrels_grades)
}
\arguments{
\item{grades}{a numeric vector of the grades.}
}
\description{
\code{num_rel} returns the number of documents retrieved.
}

