\name{Prec}
\alias{Prec}
\title{Precision at k}
\usage{
Prec(qrels_grades, grades, ranks)
}
\arguments{
  \item{qrels_grades}{a numeric vector containing the
  grades all the judged documents in the qrels file.}

  \item{grades}{a numeric vector of the grades.}

  \item{ranks}{a numeric vector containing a set of ranks
  at which the measure is calculated.}
}
\description{
\code{AP} Precision
}

