% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.qrels}
\alias{read.qrels}
\title{Read Judgments}
\usage{
read.qrels(qrelsPath, type = c("adhoc", "diversity", "triplets",
  "simulateTriplets"), opts = NULL)
}
\arguments{
\item{qrelsPath}{path to the qrels file}

\item{type}{type  of the qrels file. ('adhoc', diversity', 'triplets',
'simulateTriplets') are currently supported.}

\item{opts}{an optional parameter to provide a list of options to pass to the
parser. Options is a list containing various options as shown below:
opts=list(ties=c('random', 'aplhabetical', 'metaAP'),
          simulationType=c('subtopic'))}
}
\description{
\code{read.qrels} returns a representation of the qrels.
}
\details{
The function creates a C++ object containing various information parsed from
the qrels file.
}

