\name{read.runs}
\alias{read.runs}
\title{Sum of Vector Elements}
\usage{
  read.runs(runPaths, runids = NULL, limit = 1000,
    sort = c("trecSort", "rankSort"))
}
\description{
  \code{sum} returns the sum of all the values present in
  its arguments.
}
\details{
  This is a generic function: methods can be defined for it
  directly or via the \code{Summary} group generic.  For
  this to work properly, the arguments \code{...} should be
  unnamed, and dispatch is on the first argument.'
}

