% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adhoc.R
\name{recall}
\alias{recall}
\title{Recall}
\usage{
recall(qrels_grades, grades, rankLimit = 1000)
}
\arguments{
\item{grades}{a numeric vector of the grades.}

\item{rankLimit}{document cut off}
}
\description{
\code{recall}
}

