% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adhoc.R
\name{recip_rank}
\alias{recip_rank}
\title{Reciprocal Rank}
\usage{
recip_rank(grades, rankLimit = 1000)
}
\arguments{
\item{grades}{a numeric vector of the grades.}

\item{rankLimit}{document cut off}
}
\description{
\code{recip_rank} reciprocal rank.
}

