\name{num_rel_ret}
\alias{num_rel_ret}
\title{Number of Relevant Documents Retrieved}
\usage{
num_rel_ret(grades, rankLimit = 1000)
}
\arguments{
  \item{grades}{a numeric vector of the grades.}

  \item{rankLimit}{document cut off}
}
\description{
\code{num_rel_ret} returns the number of documents
retrieved.
}

