\name{num_ret.numeric}
\alias{num_ret.numeric}
\title{Number of Retrieved Documents}
\usage{
\method{num_ret}{numeric}(grades, rankLimit = 1000)
}
\arguments{
  \item{grades}{a numeric vector of the grades.}

  \item{rankLimit}{document cut off}
}
\description{
\code{num_ret} returns the number of documents retrieved.
}

