\name{pooling.topk}
\alias{pooling.topk}
\title{Sum of Vector Elements}
\usage{
pooling.topk(rank_matrix, k)
}
\description{
\code{sum} returns the sum of all the values present in its
arguments.
}
\details{
This is a generic function: methods can be defined for it
directly or via the \code{Summary} group generic.  For this
to work properly, the arguments \code{...} should be
unnamed, and dispatch is on the first argument.'
}

