\name{read.qrels}
\alias{read.qrels}
\title{Read Judgments}
\usage{
  read.qrels(qrelsPath, type = c("adhoc", "diversity"))
}
\description{
  \code{sum} returns the sum of all the values present in
  its arguments.
}
\details{
  This function: methods can be defined for it directly or
  via the \code{Summary} group generic.  For this to work
  properly, the arguments \code{...} should be unnamed, and
  dispatch is on the first argument.'
}

