% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.runs}
\alias{read.runs}
\title{Read System Runs}
\usage{
read.runs(runPaths, runids = NULL, limit = 1000, type = c("TREC"))
}
\arguments{
\item{runPaths}{a character vector containing a set of the paths of the runs.}

\item{runids}{a character vector containing a set of runid. Default: Uses the
filename.}

\item{limit}{number of documents to parse for each query.
Truncates document beyond this limit.}

\item{type}{indicating the type of run.
Currently only supports TREC five column format.}
}
\description{
\code{read.runs} returns a representation of a set of runs.
}
\details{
The function creates a C++ object representing a set of system runs provided
from the \code{runPaths} folder.
}

